build bugs solvd
