+++ src/lib/utils/cloudinary.ts
@@ -56,3 +56,13 @@ export const getPublicIdFromUrl = (url: string) => {
     .join("/")
     .replace(/\.[^/.]+$/, ""); // Remove file extension
 };
+
+export async function uploadFileLocally(image: File) {
+  if (image && image instanceof File) {
+    const arrayBuffer = await image.arrayBuffer();
+    const buffer = Buffer.from(arrayBuffer);
+    const base64String = buffer.toString("base64");
+    const url = await uploadImage("data:image/jpeg;base64," + base64String);
+    return url;
+  }
+}
